+++ src/HeimrichHannotContaoEntityFilterBundle.php
@@ -9,6 +9,7 @@
 namespace HeimrichHannot\EntityFilterBundle;
 
 use HeimrichHannot\EntityFilterBundle\DependencyInjection\EntityFilterExtension;
+use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class HeimrichHannotContaoEntityFilterBundle extends Bundle
@@ -16,7 +17,7 @@ class HeimrichHannotContaoEntityFilterBundle extends Bundle
     /**
      * {@inheritdoc}
      */
-    public function getContainerExtension()
+    public function getContainerExtension(): ?ExtensionInterface
     {
         return new EntityFilterExtension();
     }